refactor: restructure tests into modular files and update CI configuration
The test suite has been refactored from a single large file into multiple focused test modules under a `tests/` directory. This includes separating tests by functionality (API, commit messages, flags, git status, etc.) and adding proper test configuration files. The CI workflow and pytest configuration have been updated to reflect this new structure, improving maintainability and test organization.

# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
#
# On branch refactored-tests
#
# Changes to be committed:
# M	.github/workflows/test.yml
# M	pytest.ini
# D	test_git_commitai.py
# A	tests/__init__.py
# A	tests/conftest.py
# A	tests/test_api.py
# A	tests/test_commit_message.py
# A	tests/test_flags/__init__.py
# A	tests/test_flags/test_allow_empty.py
# A	tests/test_flags/test_amend.py
# A	tests/test_flags/test_auto_stage.py
# A	tests/test_flags/test_no_verify.py
# A	tests/test_flags/test_verbose.py
# A	tests/test_git_status.py
# A	tests/test_main.py
# A	tests/test_staged_files.py
# A	tests/test_utils.py
#
