+++ pytest.ini
@@ -1,5 +1,5 @@
 [pytest]
-testpaths = .
+testpaths = tests
 python_files = test_*.py
 python_classes = Test*
 python_functions = test_*
