+++ tests/test_flags/__init__.py
@@ -1,0 +1,1 @@
+"""Test modules for git-commitai flags."""
